requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
amqp==5.3.1
apache-superset==4.1.0
apispec==6.3.0
apsw==3.50.2.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Bottleneck==1.5.0
Brotli==1.1.0
cachelib==0.13.0
cachetools==5.5.2
cattrs==25.1.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-option-group==0.5.7
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==2.3.3
Flask-AppBuilder==4.7.0
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-Compress==1.17
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
frozenlist==1.7.0
func_timeout==4.3.5
geographiclib==2.0
geoip2==5.1.0
geopy==2.4.1
google-auth==2.40.3
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
hashids==1.3.1
holidays==0.25
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
korean-lunar-calendar==0.3.1
limits==5.4.0
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
marshmallow-sqlalchemy==1.4.2
maxminddb==2.7.0
mdurl==0.1.2
msgpack==1.0.8
msgspec==0.19.0
multidict==6.6.3
nh3==0.2.21
numba==0.60.0
numexpr==2.11.0
numpy==1.23.5
openpyxl==3.1.5
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.0.3
paramiko==3.5.1
parsedatetime==2.6
pgsanity==0.2.9
platformdirs==4.3.8
ply==3.11
polyline==2.0.2
prison==0.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-geohash==0.8.5
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
requests==2.32.4
requests-cache==1.2.1
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
selenium==4.9.1
shillelagh==1.3.5
shortid==0.1.2
simplejson==3.20.1
six==1.17.0
slack_sdk==3.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.38.3
sqlglot==25.34.1
sqlparse==0.5.3
sshtunnel==0.4.0
stdlib-list==0.11.1
tabulate==0.8.10
tkcalendar==1.6.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==2.3.7
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
